// src/elements/public/Admin/types.ts
interface Link {
  href: string;
  title: string;
  type?: string;
}

type Curies = [
  {
    name: string;
    href: string;
    templated: boolean;
  }
];

export interface FxBookmark {
  _links: {
    curies: Curies;
    self: Link;
    'fx:property_helpers': Link;
    'fx:reporting': Link;
    'fx:encode': Link;
    'fx:user': Link;
    'fx:store': Link;
    'fx:stores': Link;
    'fx:token': Link;
  };

  message: string;
}

export interface FxStore {
  _links: {
    curies: Curies;
    self: Link;
    'fx:attributes': Link;
    'fx:store_version': Link;
    'fx:users': Link;
    'fx:user_accesses': Link;
    'fx:customers': Link;
    'fx:carts': Link;
    'fx:transactions': Link;
    'fx:subscriptions': Link;
    'fx:subscription_settings': Link;
    'fx:item_categories': Link;
    'fx:taxes': Link;
    'fx:payment_method_sets': Link;
    'fx:coupons': Link;
    'fx:template_sets': Link;
    'fx:cart_templates': Link;
    'fx:cart_include_templates': Link;
    'fx:checkout_templates': Link;
    'fx:customer_portal_settings': Link;
    'fx:process_subscription_webhook': Link;
    'fx:template_configs': Link;
    'fx:payment_gateways': Link;
    'fx:payment_methods_expiring': Link;
    'fx:store_shipping_methods': Link;
    'fx:receipt_templates': Link;
    'fx:email_templates': Link;
    'fx:error_entries': Link;
    'fx:downloadables': Link;
    'fx:hosted_payment_gateways': Link;
    'fx:fraud_protections': Link;
    'fx:integrations': Link;
    'fx:native_integrations': Link;
    'fx:activate_store_monthly_url': Link;
    'fx:activate_store_yearly_url': Link;
  };
// ...
}